Automatically build and deploy html docs
Don't forget to set the SSH_PRIVATE_KEY environment variable to the
base64 encoding of the private key that should be used for
pushing.  This encoding should be generated by
```bash
$ base64 -w0 id_dsa
```
(where `id_dsa` is the private key, and `-w0` indicates no newlines
should be outputted).

The idea is that you configure a "Deploy key" on github for this
project, and feed the private key to travis directly via the
dashboard.  This way, the private key is only readable by commits on
the master branch, so not by e.g. pull requests.  And even if someone
manages to access the private key, they can only use it to push to
this repository.

When adding this commit to other repositories, don't forget to update
the remote URL in deploy_docs.sh